requirements.txt
selenium
requests
Pillow
aiohttp
imagehash
